Create a shared module and implement errorstatehandler for forms
