+++ src/modules/auth/components/signin/signin.component.html
@@ -47,7 +47,7 @@
             Sign In
           </button>
           <a
-            routerLink="/signin"
+            routerLink="/signup"
             class="link-signin"
             routerLinkActive="router-link-active"
             >Not signed up? Click to register</a
