+++ src/modules/auth/components/signin/signin.component.ts
@@ -1,25 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import {
-  FormControl,
-  FormGroupDirective,
-  NgForm,
-  Validators,
-} from '@angular/forms';
-import { ErrorStateMatcher } from '@angular/material/core';
-
-export class MyErrorStateMatcher implements ErrorStateMatcher {
-  isErrorState(
-    control: FormControl | null,
-    form: FormGroupDirective | NgForm | null
-  ): boolean {
-    const isSubmitted = form && form.submitted;
-    return !!(
-      control &&
-      control.invalid &&
-      (control.dirty || control.touched || isSubmitted)
-    );
-  }
-}
+import { FormControl, Validators } from '@angular/forms';
+import { MyErrorStateMatcher } from '../../shared/my-error-state-matcher';
+MyErrorStateMatcher;
 
 @Component({
   selector: 'app-signin',
