+++ src/modules/auth/models/user.ts
@@ -1,0 +1,4 @@
+interface user {
+  email: string;
+  password: string;
+}
