+++ src/modules/auth/shared/my-error-state-matcher.spec.ts
@@ -1,0 +1,7 @@
+import { MyErrorStateMatcher } from './my-error-state-matcher';
+
+describe('MyErrorStateMatcher', () => {
+  it('should create an instance', () => {
+    expect(new MyErrorStateMatcher()).toBeTruthy();
+  });
+});
